backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
py-clob-client>=0.19.0
supabase==2.3.4
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.31.0
pydantic==2.6.1
pydantic-settings==2.1.0

